Rename assume_role test to credentials_file
The existing assume role test doesn't test the agent assuming a role.
The test does assume a role, but only to populate a credentials file
which the agent will use to obtain credentials. The real test is the
agent using a credentials file so using a more appropriate name.
