+++ test/credentials_file/agent_config.json
@@ -6,7 +6,7 @@
         "logfile": ""
     },
     "metrics": {
-        "namespace": "AssumeRoleTest",
+        "namespace": "CredentialsFileTest",
         "append_dimensions": {
             "InstanceId": "${aws:InstanceId}"
         },
