+++ test/credentials_file/credentials_file_windows.go
@@ -4,7 +4,7 @@
 //go:build windows
 // +build windows
 
-package assume_role
+package credentials_file
 
 import (
 	"log"
